Revert "Merge branch '15.0-develop' into 15.0-1.0-develop"
This reverts commit a3d050933cd96bd7c9e7fef64d1f4e0e0f6380a6, reversing
changes made to 3449bd1b53bba977dc2d0972058cd29f44834aa0.
