+++ README.md
@@ -1,7 +1,46 @@
 
+<!-- /!\ Non OCA Context : Set here the badge of your runbot / runboat instance. -->
+[![Pre-commit Status](https://github.com/openg2p/openg2p-registry/actions/workflows/pre-commit.yml/badge.svg?branch=15.0)](https://github.com/openg2p/openg2p-registry/actions/workflows/pre-commit.yml?query=branch%3A15.0)
+[![Build Status](https://github.com/openg2p/openg2p-registry/actions/workflows/test.yml/badge.svg?branch=15.0)](https://github.com/openg2p/openg2p-registry/actions/workflows/test.yml?query=branch%3A15.0)
+[![codecov](https://codecov.io/gh/openg2p/openg2p-registry/branch/15.0/graph/badge.svg)](https://codecov.io/gh/openg2p/openg2p-registry)
+<!-- /!\ Non OCA Context : Set here the badge of your translation instance. -->
+[![Translation Status](https://translate.openspp.org/widgets/openg2p/-/svg-badge.svg)](https://translate.openspp.org/engage/openg2p/?utm_source=widget)
+
+<!-- /!\ do not modify above this line -->
+
 # OpenG2P Registry
 
-Refer to the documentation [here](https://docs.openg2p.org/modules/secure-registry).
+OpenG2P Registry is an open source project that provides a social registry that can be used on top of Odoo.
+
+<!-- /!\ do not modify below this line -->
+
+<!-- prettier-ignore-start -->
+
+[//]: # (addons)
+
+Available addons
+----------------
+addon | version | maintainers | summary
+--- | --- | --- | ---
+[g2p_bank](g2p_bank/) | 15.0.0.0.0 | [![jeremi](https://github.com/jeremi.png?size=30px)](https://github.com/jeremi) [![gonzalesedwin1123](https://github.com/gonzalesedwin1123.png?size=30px)](https://github.com/gonzalesedwin1123) [![emjay0921](https://github.com/emjay0921.png?size=30px)](https://github.com/emjay0921) | G2P Registry: Bank Details
+[g2p_registry_base](g2p_registry_base/) | 15.0.0.0.0 | [![jeremi](https://github.com/jeremi.png?size=30px)](https://github.com/jeremi) [![gonzalesedwin1123](https://github.com/gonzalesedwin1123.png?size=30px)](https://github.com/gonzalesedwin1123) | G2P Registry: Base
+[g2p_registry_group](g2p_registry_group/) | 15.0.0.0.0 | [![jeremi](https://github.com/jeremi.png?size=30px)](https://github.com/jeremi) [![gonzalesedwin1123](https://github.com/gonzalesedwin1123.png?size=30px)](https://github.com/gonzalesedwin1123) | G2P Registry: Groups
+[g2p_registry_individual](g2p_registry_individual/) | 15.0.0.0.0 | [![jeremi](https://github.com/jeremi.png?size=30px)](https://github.com/jeremi) [![gonzalesedwin1123](https://github.com/gonzalesedwin1123.png?size=30px)](https://github.com/gonzalesedwin1123) | G2P Registry: Individual
+[g2p_registry_membership](g2p_registry_membership/) | 15.0.0.0.0 | [![jeremi](https://github.com/jeremi.png?size=30px)](https://github.com/jeremi) [![gonzalesedwin1123](https://github.com/gonzalesedwin1123.png?size=30px)](https://github.com/gonzalesedwin1123) | G2P Registry: Membership
+[g2p_registry_rest_api](g2p_registry_rest_api/) | 15.0.0.0.0 | [![jeremi](https://github.com/jeremi.png?size=30px)](https://github.com/jeremi) [![gonzalesedwin1123](https://github.com/gonzalesedwin1123.png?size=30px)](https://github.com/gonzalesedwin1123) [![emjay0921](https://github.com/emjay0921.png?size=30px)](https://github.com/emjay0921) | G2P Registry: Rest API
+[g2p_registry_rest_api_extension_demo](g2p_registry_rest_api_extension_demo/) | 15.0.0.0.0 | [![jeremi](https://github.com/jeremi.png?size=30px)](https://github.com/jeremi) [![gonzalesedwin1123](https://github.com/gonzalesedwin1123.png?size=30px)](https://github.com/gonzalesedwin1123) | G2P Registry: Rest API Extension Demo
+
+[//]: # (end addons)
+
+<!-- prettier-ignore-end -->
 
 ## Licenses
+
 This repository is licensed under [MPL-2.0](LICENSE).
+
+However, each module can have a totally different license, as long as they adhere to OpenG2P
+policy. Consult each module's `__manifest__.py` file, which contains a `license` key
+that explains its license.
+
+----
+<!-- /!\ Non OCA Context : Set here the full description of your organization. -->
