--- g2p_registry_addl_info/__init__.py
@@ -1,1 +1,0 @@
-from . import models
