--- g2p_registry_addl_info/models/__init__.py
@@ -1,1 +1,0 @@
-from . import registrant
