--- g2p_registry_addl_info/readme/DESCRIPTION.rst
@@ -1,2 +1,0 @@
-This module adds support to Registry modules (registry_base, registry_individual and registry_group)
-to accept a new field called additional_g2p_info, which stores data as json and displays it as form.
