--- g2p_registry_addl_info_rest_api/README.rst
@@ -1,54 +1,0 @@
-=============================
-G2P Registry: Additional Info
-=============================
-
-.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! This file is generated by oca-gen-addon-readme !!
-   !! changes will be overwritten.                   !!
-   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
-.. |badge1| image:: https://img.shields.io/badge/maturity-Alpha-red.png
-    :target: https://odoo-community.org/page/development-status
-    :alt: Alpha
-.. |badge2| image:: https://img.shields.io/badge/github-OpenG2P%2Fopeng2p--registry-lightgray.png?logo=github
-    :target: https://github.com/Ope"G2P Registry: Additional Info"nG2P/openg2p-registry/tree/15.0-develop/g2p_registry_addl_info_rest_api
-    :alt: OpenG2P/openg2p-registry
-
-|badge1| |badge2| 
-
-This module adds REST API support to "G2P Registry: Additional Info"
-
-.. IMPORTANT::
-   This is an alpha version, the data model and design can change at any time without warning.
-   Only for development or testing purpose, do not use in production.
-   `More details on development status <https://odoo-community.org/page/development-status>`_
-
-**Table of contents**
-
-.. contents::
-   :local:
-
-Bug Tracker
-===========
-
-Bugs are tracked on `GitHub Issues <https://github.com/OpenG2P/openg2p-registry/issues>`_.
-In case of trouble, please check there if your issue has already been reported.
-If you spotted it first, help us smashing it by providing a detailed and welcomed
-`feedback <https://github.com/OpenG2P/openg2p-registry/issues/new?body=module:%20g2p_registry_addl_info_rest_api%0Aversion:%2015.0-develop%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
-
-Do not contact contributors directly about support or help with technical issues.
-
-Credits
-=======
-
-Authors
-~~~~~~~
-
-* OpenG2P
-
-Maintainers
-~~~~~~~~~~~
-
-This module is part of the `OpenG2P/openg2p-registry <https://github.com/OpenG2P/openg2p-registry/tree/15.0-develop/g2p_registry_addl_info_rest_api>`_ project on GitHub.
-
-You are welcome to contribute.
