--- g2p_registry_addl_info_rest_api/__manifest__.py
@@ -1,21 +1,0 @@
-{
-    "name": "G2P Registry: Additional Info REST API",
-    "category": "G2P",
-    "version": "15.0.1.1.0",
-    "sequence": 1,
-    "author": "OpenG2P",
-    "website": "https://github.com/openg2p/openg2p-registry",
-    "license": "Other OSI approved licence",
-    "development_status": "Alpha",
-    "depends": ["g2p_registry_rest_api", "g2p_registry_addl_info"],
-    "data": [],
-    "assets": {
-        "web.assets_backend": [],
-        "web.assets_qweb": [],
-    },
-    "demo": [],
-    "images": [],
-    "application": False,
-    "installable": True,
-    "auto_install": False,
-}
