--- g2p_registry_addl_info_rest_api/models/__init__.py
@@ -1,2 +1,0 @@
-from . import registrant
-from . import group_membership
