--- g2p_registry_addl_info_rest_api/readme/DESCRIPTION.rst
@@ -1,1 +1,0 @@
-This module adds REST API support to "G2P Registry: Additional Info"
