--- g2p_registry_addl_info_rest_api/services/__init__.py
@@ -1,2 +1,0 @@
-from . import process_individual_mixin
-from . import process_group_mixin
