--- g2p_registry_addl_info_rest_api/services/process_group_mixin.py
@@ -1,11 +1,0 @@
-from odoo.addons.component.core import AbstractComponent
-
-
-class ProcessGroupMixin(AbstractComponent):
-    _inherit = "process_group.rest.mixin"
-
-    def _process_group(self, group_info):
-        res = super(ProcessGroupMixin, self)._process_group(group_info)
-        if group_info.dict().get("additional_g2p_info", None):
-            res["additional_g2p_info"] = group_info.additional_g2p_info
-        return res
