+++ g2p_registry_rest_api/__manifest__.py
@@ -2,7 +2,7 @@
 {
     "name": "G2P Registry: Rest API",
     "category": "G2P",
-    "version": "15.0.1.1.0",
+    "version": "15.0.1.0.1",
     "sequence": 1,
     "author": "OpenG2P",
     "website": "https://github.com/openg2p/openg2p-registry",
