+++ g2p_registry_rest_api/models/group_membership.py
@@ -21,7 +21,6 @@ class GroupMembersInfoOut(NaiveOrmModel):
 class GroupMembersInfoIn(NaiveOrmModel):
     name: str
     given_name: str = None
-    addl_name: str = None
     family_name: str = None
     ids: List[RegistrantIDIn] = None
     registration_date: date = None
