--- g2p_registry_rest_api/models/individual_search_param.py
@@ -1,7 +1,0 @@
-from extendable_pydantic import ExtendableModelMeta
-from pydantic import BaseModel
-
-
-class IndividualSearchParam(BaseModel, metaclass=ExtendableModelMeta):
-    id: int = None
-    name: str = None
