+++ g2p_registry_rest_api/models/registrant.py
@@ -55,11 +55,3 @@ class RegistrantInfoIn(NaiveOrmModel):
     phone_numbers: List[PhoneNumberIn] = None
     email: str = None
     address: str = None
-
-
-class RegistrantUpdateIDIn(RegistrantIDIn):
-    partner_id: int
-
-
-class RegistrantUpdateIDOut(RegistrantIDOut):
-    partner_id: int
