+++ g2p_registry_rest_api/services/__init__.py
@@ -1,4 +1,1 @@
-from . import process_individual_mixin
-from . import process_group_mixin
 from . import group
-from . import individual
