--- setup/g2p_registry_addl_info/odoo/addons/g2p_registry_addl_info
@@ -1,1 +1,0 @@
-../../../../g2p_registry_addl_info
