--- setup/g2p_registry_addl_info/setup.py
@@ -1,6 +1,0 @@
-import setuptools
-
-setuptools.setup(
-    setup_requires=['setuptools-odoo'],
-    odoo_addon=True,
-)
