--- setup/g2p_registry_addl_info_rest_api/odoo/addons/g2p_registry_addl_info_rest_api
@@ -1,1 +1,0 @@
-../../../../g2p_registry_addl_info_rest_api
